+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Changelog
  * `CompositeRepository` now sets the correct path for returned resources
  * moved `Selector` to `Puli\Repository\Selector` namespace
  * removed tagging, which is out of scope of this package
+ * made `AbstractResource` properties private
 
 * 1.0.0-alpha4 (2014-12-03)